Add map.h
Variadic template that we will use to track all the
items on the recursion tree of bytewise exploration.

Implements a method to append a map to another, and
one to prefix all the paths in a map with a new step,
which we will use to join trees during the recursive
bytewise exploration.