Add bytewise/proxy
Forwards all its calls to the bytewise interface of an
object. Bytewise macro now friends the proxy. It will
be sufficient to friend the proxy with all the classes
that need to access a bytewise interface, and we are
good to go: there will be no need to add a verbose
list of friends directly to the macro.