Add bytewise/mask.h
Variadic template that we will use to track all the
items on the recursion tree of bytewise exploration.

Implements a method to append a mask to another, and
one to shift all the ranges in a map by some value,
which we will use to join masks during the recursive
bytewise exploration.