Update scanners/arithmetic.h
Adds two new features: empty and writable. Empty means
that there is no bytewise field in the tree of
recursion. Writable means that all the elements in
the tree of recursion can be written (which means:
none of them are const or members of const classes).

We will use writable in the iterator to determine if
an object can be accessed in write mode.