Add bytewise/serialize.h
First sketch of serialization. Uses scanners to
appropriately allocate itself (at compile time if
possible), then uses visitors to actually serialize
data.

When visiting variable-size buffers, it prepends to
each chunk of data an uint32_t value with the size
of the variable-size buffer. This last mechanism
should be improved: often times four bytes is just
a waste of space.