Update data/variant
Adds copy and move assignment to variants. It also
includes assignment to variants with a different
type_id. Note how disabling copy and move constructors
in one of the types in a variant also means disabling
its assignment. In fact, in case of a copy that
affects the type_id, the old type gets destructed and
a new one needs to be created.