Add utils/template/is_callable.h
Sfinae-based struct that tells if a functor is
callable with a given argument. Still to be extended
to multiple arguments.

Offers two values: value just says if it is callable
with that value. Direct says if it is callable
without any cast, i.e., the functor explicitly
accepts that value.