Update utils/template/is_callable.h
Const checks are now performed only if non-const
checks fail. So if the functor is a lambda with an
auto && parameter, is_callable tries to only
specialize it with const if it does not accept
a non-const reference.

This way, implementations that have an auto &&
argument but use it as non-const will never be
specialized to a const auto argument, thus failing
to compile.