Update async/promise
When a lambda is provided to then that returns a
promise, a temporary promise is generated as a member
of the callback object. The member chain promise is
then returned from the then method, and the user can
proceed to add thens. When the actual promise is
returned from the callback, the temporary promise
becomes an alias to that promise. All the operations
on it will be forwarded to the other promise.

Still needs:
 - Garbage collecting.
 - Some mechanism that prevents resolving temporary
   promises.