Add network/pool/queue
First sketch of queue. Works on kqueues in OSX.
Linux still to be implemented (with epolls). Offers
add() and remove methods() to appropriately add and
remove descriptors. Offers select() that awaits for
events up to a specified timeout, and returns the
number of available sockets.

The [] operator allows to loop through the events
received.