Add data/heap
Basic implementation of heap. Includes appropriate
on-the-go construction and destruction using
optionals.

Includes fix to optionals that prevents calling
the parametric constructor of optional_base when
copying an optional that can cast to the type it
stores (as in the case of arithmetics that can be
constructed from the bool casting of optional).