Update network/pool/pool, async/context
Second round of pool optimization. When calling send
on a pool connection, the pool will first try to
directly fit everything into the socket's buffer.
Only if it fails to do so, the send operation is
added to the pool.

Additional bugfix: when a context returns from a
round of run and the promise with which it leaves
is already resolved, the context will recall itself
directly, adding to the stack. If too many promises
are resolved on the spot, this would cause a stack
overflow. Now the exit type has a way of signaling
if the run round from which it was returned should
be immediately run again.