Update network/dns/message
Adds answer method to packet that turns it into a
response message and adds the appropriate field.

The example in main shows how this can be used to
make a simple authoritative server that points any
question to 127.0.0.1.

There must be some bug in the visit() method on
const variants, which is why we have to copy the
query before visiting it. Fix it soon.